+++ GW2EIEvtcParser/EIData/CombatReplay/Decorations/FacingDecoration.cs
@@ -4,7 +4,7 @@ namespace GW2EIEvtcParser.EIData
 {
     internal class FacingDecoration : GenericAttachedDecoration
     {
-        public List<int> Angles { get; } = new List<int>();
+        public List<float> Angles { get; } = new List<float>();
 
         public FacingDecoration((int start, int end) lifespan, AgentConnector connector, List<Point3D> facings) : base(lifespan, connector)
         {
